refactor: move dependency-helper location
