+++ apps/web/src/lib/variable-models/_helpers/dependency-helper.ts
